Code/requirements/requirements_qwen.txt
absl-py==2.1.0
accelerate==1.8.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
av==13.0.0
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.2
decord==0.6.0
deepspeed==0.15.4
dill==0.3.8
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.16.1
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.9.0
grpcio==1.66.1
hjson==3.1.0
huggingface-hub==0.31.2
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
modelscope==1.23.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
ninja==1.11.1.1
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pdf2image==1.17.0
peft==0.15.1
pillow==10.4.0
propcache==0.2.0
protobuf==5.28.1
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qwen-vl-utils==0.0.8
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
six==1.16.0
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
tqdm==4.66.5
transformers==4.53.2
transformers-stream-generator==0.0.4
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
xxhash==3.5.0
yarl==1.16.0
